# internal/api/openapi.yaml
openapi: 3.1.0
info:
  title: cloudflared-manager API
  description: |
    无头（headless）cloudflared **TOKEN 模式** 隧道管理器。守护进程 `cfdmgrd` 把每个
    cloudflared 连接器（connector）放进**独立的外部子进程**（`cloudflared tunnel run --token …`），
    通过 HTTP/JSON 暴露隧道 CRUD、生命周期、历史曲线、告警、日志、导入导出、系统监控、
    cloudflared 二进制版本管理等管理面。前端 SPA 已 embed 进二进制，生产环境同域。

    关键架构事实（落到 API 字段命名、状态码、错误形态都需要先看一眼）：

    - **外部子进程模型**：每个运行中的隧道都是 `cloudflared tunnel run` 子进程，由
      `internal/process.Worker` 监管。父进程通过 `--token`（环境变量 `TUNNEL_TOKEN`）注入
      连接器令牌；ingress / public-hostname / origin 等参数全部在 Cloudflare Zero Trust
      控制台维护，**不在本模型内**。
    - **reload = 重启**：cloudflared 没有针对单连接器参数的 in-place 热重载语义，
      `POST .../reload` 的实现就是 `stop()` 再 `start()`。
    - **token 高度敏感**：所有返回 `ConfigEnvelope` 的接口里 `config.token` 一律被剥除（空串），
      仅用布尔 `has_token` 表明是否已存储；掩码预览经 `GET /configs/{id}/token` 获取。
      **唯一例外** 是 `GET /configs/{id}/raw`（原样回写磁盘 YAML，按设计包含 token，属敏感导出）。
    - **流量曲线语义特殊**：cloudflared **没有按隧道的字节计数器**。
      `GET /metrics/{id}/traffic` 返回的 `points[].in` = 请求数增量、`out` = 错误数增量、
      `conns` = HA 连接数 gauge（**不是字节**），均来自抓取 cloudflared 自带的 `--metrics`
      Prometheus 端点。
    - **配置数据模型**：业务层（请求/响应里的 `config` 字段）= `TunnelConfigV1`，**全 camelCase**
      （edge/reliability/logging/identity/advancedEnvOverrides/binaryVersion）。
    - **快照 / 事件 / 告警 / 流量 / 系统监控**：`Snapshot`、`/metrics/*`、`/alerts/*`、`/system/*`、
      WebSocket `/events` 帧统一走 **snake_case**（id / inst_id / started_at / metrics_port …）。
...

...

...
  version: 2.0.0
servers:
...
security:
  - BearerAuth: []
...
paths:
...
  /api/v1/version/check:
    get:
      tags: [system]
      summary: 检查是否有新版本可自更新
      description: 对比运行版本与 GitHub 最新 release，并报告部署方式与一键自更新能力。`?force=1` 跳过约 1 小时缓存。
      parameters:
        - { name: force, in: query, required: false, schema: { type: string }, description: "1 / refresh=1 跳过缓存" }
      responses:
        "200":
          description: 检查结果
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VersionCheck" }
  /api/v1/system/update:
    post:
      tags: [system]
      summary: 触发分离式自更新并重启
      description: 立即返回 202，守护进程随后被替换并重启；客户端应轮询 /health 与 /version 直到版本变化。`?force=1` 强制重装最新版。
      parameters:
        - { name: force, in: query, required: false, schema: { type: string }, description: "1 强制重装当前最新版" }
      responses:
        "202":
          description: 更新已开始
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: updating }
                  from: { type: string }
                  to: { type: string }
                  message: { type: string }
        "400": { $ref: "#/components/responses/BadRequest", description: 部署方式不支持自更新 }
        "403": { $ref: "#/components/responses/Error", description: Web 端自更新被禁用（CFDM_SELF_UPDATE_ENABLED=false） }
        "409": { $ref: "#/components/responses/Conflict", description: "已是最新版本（可加 ?force=1）" }
        "502": { $ref: "#/components/responses/Error", description: 无法获取最新版本（上游失败） }
...
  /api/v1/ui/branding:
    get:
      tags: [system]
      summary: 读取生效的 UI 品牌（公开，无需鉴权）
      description: 返回侧边栏 / 登录页 / 浏览器标题使用的品牌；空字段已用默认值填充。登录页与首屏 <title> 在鉴权前即可渲染，故公开。
      security: []
      responses:
        "200":
          description: 生效品牌
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Branding" }
    put:
      tags: [system]
      summary: 更新 UI 品牌（需鉴权）
      description: 省略的字段保留当前存储值，显式空串重置为默认。品牌持久化在数据目录 meta.json，跨设备 / 清缓存后依然生效，首屏由服务端注入实现零闪。
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Branding" }
      responses:
        "200":
          description: 生效品牌
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Branding" }
  /api/v1/configs:
    get:
      tags: [configs]
      summary: 列出所有隧道配置快照
      responses:
        "200":
          description: 快照列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/Snapshot" }
    post:
      tags: [configs]
      summary: 新建隧道配置
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id, config]
              properties:
                id: { type: string, description: 配置 id（文件名干）}
                config: { $ref: "#/components/schemas/TunnelConfigV1" }
                cfdmgr: { $ref: "#/components/schemas/MgrMeta" }
      responses:
        "201":
          description: 已创建
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ConfigEnvelope" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/configs/reorder:
    post:
      tags: [configs]
      summary: 持久化前端展示排序
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [order]
              properties:
                order:
                  type: array
                  items: { type: string }
                  description: 按期望顺序排列的配置 id
      responses:
        "204": { description: 已保存 }
  /api/v1/configs/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string } }
    get:
      tags: [configs]
      summary: 获取单个配置（含 TunnelConfigV1，token 已剥除）
      responses:
        "200":
          description: 配置信封
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ConfigEnvelope" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [configs]
      summary: 整体替换配置体
      description: 提交空 / 省略 token 表示保留已存储的 token；仅非空 token 才会替换密钥。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [config]
              properties:
                config: { $ref: "#/components/schemas/TunnelConfigV1" }
                cfdmgr: { $ref: "#/components/schemas/MgrMeta" }
      responses:
        "200":
          description: 已更新
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ConfigEnvelope" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [configs]
      summary: RFC 7396 合并补丁（merge-patch）应用于 TunnelConfigV1
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema:
              type: object
              description: TunnelConfigV1 的部分字段；null 值删除对应键
      responses:
        "200":
          description: 已更新
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ConfigEnvelope" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [configs]
      summary: 停止并删除隧道配置
      responses:
        "204": { description: 已删除 }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/configs/{id}/duplicate:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string } }
    post:
      tags: [configs]
      summary: 复制配置到新 id（token 不会被复制）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [new_id]
              properties:
                new_id: { type: string }
      responses:
        "201":
          description: 已复制
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ConfigEnvelope" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/configs/{id}/raw:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string } }
    get:
      tags: [configs]
      summary: 获取磁盘原始 YAML（敏感：按设计包含 token）
      responses:
        "200":
          description: 原始 cloudflared YAML
          content:
            text/yaml:
              schema: { type: string }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [configs]
      summary: 用原始 cloudflared YAML 替换磁盘文件
      requestBody:
        required: true
        content:
          text/yaml:
            schema: { type: string }
      responses:
        "200":
          description: 已写入
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ConfigEnvelope" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/configs/{id}/token:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string } }
    get:
      tags: [configs]
      summary: 获取连接器 token 的掩码预览（永不返回明文）
      responses:
        "200":
          description: 掩码信息
          content:
            application/json:
              schema:
                type: object
                properties:
                  has_token: { type: boolean }
                  masked: { type: string, description: "首 4 + 圆点 + 末 4 的不可逆预览" }
                  length: { type: integer }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/configs/{id}/start:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string } }
    post:
      tags: [configs]
      summary: 启动隧道实例
      responses:
        "200":
          description: 启动后的快照
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Snapshot" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict", description: 已在运行 }
  /api/v1/configs/{id}/stop:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string } }
    post:
      tags: [configs]
      summary: 停止隧道实例
      responses:
        "200":
          description: 停止后的快照
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Snapshot" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/configs/{id}/reload:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string } }
    post:
      tags: [configs]
      summary: 重载隧道实例（实现为 stop + start）
      responses:
        "200":
          description: 重载后的快照
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Snapshot" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/configs/{id}/status:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string } }
    get:
      tags: [configs]
      summary: 获取隧道实例运行状态快照
      responses:
        "200":
          description: 快照
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Snapshot" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /api/v1/alerts/events:
    get:
      tags: [metrics]
      summary: 列出告警事件
      parameters:
        - { name: state, in: query, required: false, schema: { type: string, enum: [firing, resolved] } }
        - { name: from, in: query, required: false, schema: { type: integer, format: int64 } }
        - { name: to, in: query, required: false, schema: { type: integer, format: int64 } }
      responses:
        "200":
          description: 事件列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/AlertEvent" }
  /api/v1/alerts:
    get:
      tags: [metrics]
      summary: 列出告警规则
      responses:
        "200":
          description: 规则列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/AlertRule" }
    post:
      tags: [metrics]
      summary: 创建告警规则（省略 id 时自动生成）
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AlertRule" }
      responses:
        "201":
          description: 已创建
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AlertRule" }
        "400": { $ref: "#/components/responses/BadRequest", description: "metric 须为 conns|requests_rate|errors_rate（traffic_in_rate/traffic_out_rate 为遗留别名），op 须为 >|>=|<|<=" }
  /api/v1/alerts/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string } }
    get:
      tags: [metrics]
      summary: 获取单条告警规则
      responses:
        "200":
          description: 规则
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AlertRule" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [metrics]
      summary: 替换告警规则（id 取自路径）
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AlertRule" }
      responses:
        "200":
          description: 已更新
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AlertRule" }
        "400": { $ref: "#/components/responses/BadRequest" }
    delete:
      tags: [metrics]
      summary: 删除告警规则
      responses:
        "204": { description: 已删除 }
  /api/v1/validate:
    post:
      tags: [configs]
      summary: 校验隧道配置（不持久化）
      description: "Content-Type 为 application/json 时按 TunnelConfigV1 解析；其它 Content-Type 按原始 cloudflared YAML 解析。"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TunnelConfigV1" }
          text/yaml:
            schema: { type: string }
      responses:
        "200":
          description: 校验结果（语法/语义错误也以 200 + valid=false 返回）
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid: { type: boolean }
                  errors: { type: array, items: { type: string } }
                  warnings: { type: array, items: { type: string } }
  /api/v1/configs/{id}/logs:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string } }
    get:
      tags: [logs]
      summary: 查询隧道日志最近若干行
      parameters:
        - { name: lines, in: query, required: false, schema: { type: integer, default: 200 } }
        - { name: level, in: query, required: false, schema: { type: string } }
        - { name: keyword, in: query, required: false, schema: { type: string } }
        - { name: since, in: query, required: false, schema: { type: integer, format: int64 } }
      responses:
        "200":
          description: 日志行
          content:
            application/json:
              schema:
                type: object
                properties:
                  lines: { type: array, items: { type: string } }
                  next_offset: { type: integer, format: int64 }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [logs]
      summary: 清空日志视图（设置"已清空"水位线，不删除文件）
      responses:
        "204": { description: 已设置水位线 }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/configs/{id}/logs/files:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string } }
    get:
      tags: [logs]
      summary: 列出本实例日志文件（含轮转副本）
      responses:
        "200":
          description: 文件列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        path: { type: string }
                        rotated_at: { type: string, format: date-time }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/configs/{id}/logs/tail:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string } }
      - { name: token, in: query, required: false, schema: { type: string }, description: WebSocket 鉴权令牌 }
    get:
      tags: [logs]
      summary: 实时跟踪日志（WebSocket 升级，原始文本行）
      description: "升级为 WebSocket；每帧为 JSON `{\"line\":\"<日志行>\"}`。结构化日志请用 /logs/stream。"
      responses:
        "101": { description: 协议切换为 WebSocket }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/configs/{id}/logs/stream:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string } }
      - { name: token, in: query, required: false, schema: { type: string }, description: WebSocket 鉴权令牌 }
      - { name: level, in: query, required: false, schema: { type: string, enum: [debug, info, warn, error, fatal] }, description: 最低级别过滤 }
      - { name: keyword, in: query, required: false, schema: { type: string }, description: 子串过滤（匹配 message+raw，大小写不敏感） }
      - { name: conn_index, in: query, required: false, schema: { type: integer }, description: 仅某条 edge 连接的日志 }
      - { name: backlog, in: query, required: false, schema: { type: integer, default: 500 }, description: 连接时回放的历史行数 }
    get:
      tags: [logs]
      summary: 结构化实时日志流（WebSocket 升级）
      description: |
        升级为 WebSocket，订阅子进程结构化日志。每帧为 JSON `{"entries":[LogEntry,...]}`：
        连接时先发一帧历史（按 backlog + 过滤 + 清空水位），之后每条新行一帧（长度 1）。
        前端按 `seq` 去重（订阅先于快照，边界可能出现一条重复）。
      responses:
        "101": { description: "协议切换为 WebSocket（每帧为 entries 数组）" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /api/v1/configs/{id}/projection:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string } }
    get:
      tags: [configs]
      summary: 运行参数投影（cfdflags 投影出的真实 TUNNEL_* env / argv）
      description: "展示该实例实际以什么参数运行。TUNNEL_TOKEN 始终脱敏，绝不回明文。"
      responses:
        "200":
          description: 投影结果
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Projection" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/events:
    get:
      tags: [events]
      summary: 事件推送（WebSocket 升级）
      description: "升级为 WebSocket，推送 Event 帧。事件类型包含 config.changed / config.deleted / instance.state / instance.error / alert 等。"
      parameters:
        - { name: token, in: query, required: false, schema: { type: string }, description: WebSocket 鉴权令牌 }
        - { name: since, in: query, required: false, schema: { type: integer, format: int64 }, description: 回放 seq 大于该值的历史事件 }
        - { name: types, in: query, required: false, schema: { type: string }, description: 逗号分隔的事件类型过滤 }
        - { name: config_ids, in: query, required: false, schema: { type: string }, description: 逗号分隔的配置 id 过滤 }
      responses:
        "101": { description: 协议切换为 WebSocket（每帧为 Event） }
  /api/v1/import/file:
    post:
      tags: [importexport]
      summary: 导入单个 .yaml/.yml 文件
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
                id: { type: string, description: 可选，覆盖以文件名为准的 id }
      responses:
        "200":
          description: 已导入
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ConfigEnvelope" }
        "400": { $ref: "#/components/responses/BadRequest" }
  /api/v1/import/url:
    post:
      tags: [importexport]
      summary: 从 URL 导入配置（仅允许公网 http/https）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url]
              properties:
                url: { type: string }
                id: { type: string }
      responses:
        "200":
          description: 已导入
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ConfigEnvelope" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "502": { $ref: "#/components/responses/Error", description: 下载失败 }
  /api/v1/import/text:
    post:
      tags: [importexport]
      summary: 从文本（cloudflared YAML）导入配置
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id, text]
              properties:
                id: { type: string }
                text: { type: string, description: cloudflared YAML 文本 }
      responses:
        "200":
          description: 已导入
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ConfigEnvelope" }
        "400": { $ref: "#/components/responses/BadRequest" }
  /api/v1/import/zip:
    post:
      tags: [importexport]
      summary: 导入 /export/all 生成的 zip 备份
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary, description: 由 /export/all 导出的 zip }
      responses:
        "200":
          description: 已导入
          content:
            application/json:
              schema:
                type: object
                properties:
                  imported: { type: array, items: { type: string } }
        "400": { $ref: "#/components/responses/BadRequest" }
  /api/v1/configs/{id}/export:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string } }
    get:
      tags: [importexport]
      summary: 导出单个配置 YAML（下载）
      responses:
        "200":
          description: YAML 下载
          content:
            text/yaml:
              schema: { type: string }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /api/v1/system/cpu:
    get:
      tags: [system]
      summary: CPU 指标
      parameters:
        - { name: window, in: query, required: false, schema: { type: string, default: 200ms }, description: 采样窗口（最大 5s）}
      responses:
        "200":
          description: CPU 指标
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
...
  /api/v1/system/disk:
    get:
      tags: [system]
      summary: 磁盘使用（数据目录 + 根）
      parameters:
        - { name: paths, in: query, required: false, schema: { type: string }, description: 逗号分隔的额外绝对路径 }
      responses:
        "200":
          description: 磁盘项列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: { type: array, items: { type: object, additionalProperties: true } }
...
  /api/v1/binaries/available:
    get:
      tags: [binaries]
      summary: 列出 GitHub 上可下载的 release
      responses:
        "200":
          description: 可用 release 列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/AvailableRelease" }
        "502": { $ref: "#/components/responses/Error", description: 拉取 release 失败 }
  /api/v1/binaries/install:
    post:
      tags: [binaries]
      summary: 下载并安装一个 cloudflared 版本
      description: "version 允许 `latest`；省略时回退到 CFDM_CLOUDFLARED_DEFAULT_VERSION（默认 latest）。安装后不自动激活。"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                version: { type: string }
      responses:
        "201":
          description: 已安装
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VersionMeta" }
        "400": { $ref: "#/components/responses/BadRequest", description: 非法 version tag }
        "502": { $ref: "#/components/responses/Error", description: 安装 / 下载失败 }
  /api/v1/binaries/{version}/activate:
    parameters:
      - { name: version, in: path, required: true, schema: { type: string } }
    post:
      tags: [binaries]
      summary: 将某版本设为当前激活版本
      responses:
        "200":
          description: 已激活
          content:
            application/json:
              schema:
                type: object
                properties:
                  version: { type: string }
                  active: { type: boolean }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound", description: 版本未安装 }
  /api/v1/binaries/{version}:
    parameters:
      - { name: version, in: path, required: true, schema: { type: string } }
    delete:
      tags: [binaries]
      summary: 删除已安装版本
      responses:
        "204": { description: 已删除 }
        "404": { $ref: "#/components/responses/NotFound", description: 版本未安装 }
        "409": { $ref: "#/components/responses/Conflict", description: 当前激活版本不可删除 }

  # ── Cloudflare 账号直连集成 ───────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: "Authorization: Bearer <CFDM_API_TOKEN>。WebSocket 接口可改用 ?token= 查询参数。"
  responses:
    BadRequest:
      description: 请求非法
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    Conflict:
      description: 状态冲突
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Error:
      description: 错误
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
...
    Snapshot:
      type: object
      description: 单个隧道实例的运行时状态视图（snake_case）
      properties:
        id: { type: string }
        name: { type: string }
        path: { type: string, description: 配置文件路径（.yaml）}
        log_path: { type: string }
        state: { type: string, enum: [stopped, starting, started, stopping] }
        last_error: { type: string }
        started_at: { type: string, format: date-time }
        stopped_at: { type: string, format: date-time }
        binary_version: { type: string, description: 该实例当前使用的 cloudflared 版本 }
        pid: { type: integer, description: 子进程 pid，0 表示未运行 }
        metrics_port: { type: integer, description: 分配给该实例的 metrics 端口 }
    MgrMeta:
      type: object
      description: 管理器级元数据（存于 meta.json，不在隧道 YAML 内）
      properties:
        name: { type: string, description: 显示名 }
        manualStart: { type: boolean, description: 是否禁用守护进程自启动 }
    ConfigEnvelope:
      type: object
      description: "Snapshot 字段 + config（TunnelConfigV1，token 已剥除）+ cfdmgr + has_token。"
      allOf:
        - $ref: "#/components/schemas/Snapshot"
        - type: object
          properties:
            config: { $ref: "#/components/schemas/TunnelConfigV1" }
            cfdmgr: { $ref: "#/components/schemas/MgrMeta" }
            has_token:
              type: boolean
              description: 是否已存储 token（不泄露明文）
...
    VersionCheck:
      type: object
      description: "GET /version/check 的响应；获取最新版失败时仅含 current / *_enabled / has_update=false / can_self_update=false / reason / check_error。"
      properties:
        current: { type: string }
        latest: { type: string }
        deployment_mode: { type: string }
        self_update_enabled: { type: boolean }
        has_update: { type: boolean }
        can_self_update: { type: boolean }
        reason: { type: string }
        changelog: { type: string }
        html_url: { type: string }
        published_at: { type: string, format: date-time }
        check_error: { type: string }
    Branding:
      type: object
      description: "UI 品牌（snake_case）。GET 返回生效值（空字段已填默认）；PUT 时省略字段保留原值、空串重置为默认。"
      properties:
        app_name: { type: string, maxLength: 40, description: "品牌名（侧边栏 / 登录页顶部），默认「Cloudflared 隧道管理器」" }
        app_subtitle: { type: string, maxLength: 60, description: "副标题（登录页小字），默认「Cloudflare Tunnel 管理面板」" }
        html_title: { type: string, maxLength: 120, description: "浏览器标签 <title>，默认「Cloudflared 隧道管理器 · Cloudflare Tunnel 管理面板」" }

    # ── Cloudflare 集成 schemas ──────────────────────────────────────────
...